requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-taggit
django-debug-toolbar
django-filter
django-environ
celery
pyotp
requests
kavenegar
Pillow
faker